+++ lib/markly/flags.rb
@@ -26,12 +26,12 @@ module Markly
 	# Parse language prefixes on inline code spans, e.g. ruby:`Object.new`.
 	INLINE_CODE_INFO = 1 << 19
 	# Allow consistently indented HTML content to continue across blank lines.
-	INDENTED_HTML_BLOCKS = 1 << 20
+	HTML_BLOCK_BLANK_LINES = 1 << 20
 	
 	PARSE_FLAGS = {
 		front_matter: FRONT_MATTER,
 		inline_code_info: INLINE_CODE_INFO,
-		indented_html_blocks: INDENTED_HTML_BLOCKS,
+		html_block_blank_lines: HTML_BLOCK_BLANK_LINES,
 		validate_utf8: VALIDATE_UTF8,
 		smart_quotes: SMART,
 		liberal_html_tags: LIBERAL_HTML_TAG,